refactor(ui): merge analytics tab into dashboard homepage
Consolidate analytics into the dashboard by removing the separate /analytics route and Analytics.tsx page. All analytics content (outcome signals, decision types, trace quality trends, conflict analytics, agent scorecard) now renders directly on the dashboard below existing metric cards. Updated display limits: top-5 instead of top-10 for decision types and agent pairs, top-5 for agent scorecard. Limited recent decisions list to 5. Added period selector (7d/30d/90d) to dashboard header to control trend data.
