+++ ui/src/components/Layout.tsx
@@ -13,7 +13,6 @@ import {
   Search,
   Download,
   Shield,
-  BarChart3,
   LogOut,
   Menu,
   Moon,
@@ -48,7 +47,6 @@ const navItems = [
   { to: "/agents", label: "Agents", icon: Users },
   { to: "/conflicts", label: "Conflicts", icon: AlertTriangle },
   { to: "/search", label: "Search", icon: Search },
-  { to: "/analytics", label: "Analytics", icon: BarChart3 },
   { to: "/export", label: "Export", icon: Download },
   { to: "/grants", label: "Grants", icon: Shield },
 ];
